chore: split DevicesTab into DeviceIconSet, DevicesThisDevice, DevicesNearby, DevicesSyncOptions
